Original commit message from CVS:
Merge from HEAD to BRANCH-GOBJECT1 at 200106241

+++ gst/elements/gstdisksink.c
@@ -216,7 +216,7 @@ static void
 gst_disksink_chain (GstPad *pad, GstBuffer *buf) 
 {
   GstDiskSink *disksink;
-  guint16 bytes_written = 0;
+  gint bytes_written = 0;
 
   g_return_if_fail (pad != NULL);
   g_return_if_fail (GST_IS_PAD (pad));
